Fix use-after-free in blockwise coap JNI
Summary:
We have a use-after-free race between canceling a blockwise request and completing the blockwise request. Ensuring that all frees occur on the GG thread will help this situation, but will not fix it entirely. Further work is needed to add an additional layer of indirection that allows us to determine whether a pointer has been freed when receiving that pointer from java-land.

Test Plan: has verified that the reproduction steps he found previously no longer work, but a bit of regression testing before updating the GG lib in the MFA is warranted.