FC-5833 CoAP does not work after phone reboot
_The reported issue_

After rebooting an iOS 14 phone, syncing doesn't work for GG devices paired with the Fitbit iOS app running on that phone. In fact no communication can be made between phone and tracker. The app must be restarted or BT toggled to resume communication with tracker.

_Root cause analysis_

As Gilles and others mentioned on Jira, the problem seems to be somewhere between TLS and CoAP, on the phone side.

Normally, the Coap endpoint sets itself as a data sink for the user side of the DTLS protocol. However, in some cases, the DTLS handshake happens very fast, before the Coap layer has the chance to set itself as the data sink for the DTLS layer. If there's no data sink set, the DTLS layer buffers the first packet and rejects all other incoming traffic until the stored packet is consumed. Now, the problem is that when the Coap endpoint finally sets itself as the data sink, the DTLS layer takes no action (like trying to pump the stored packet). The DTLS layer remains stuck forever because there's no one asking for the buffered packet.

As to why this pre-existing issue is visible only on iOS 14 devices, it may be because Apple imposed further restrictions on the number of threads/queues an application is allowed to create. I am seeing the the GG runloop now runs in an overcommitted queue. Moreover, when iOS boots, the Fitbit app has more limited CPU time (because iOS itself uses CPU to start up various services). This translates to the situation where the Coap layer sets itself as a data sink for the DTLS layer a bit later than usual.

This issue likely affects Android and FW systems as well, if the data sink is set a bit later.

_Easy repro_

I've made a minor change in GG to reproduce the issue consistently, with GGHost, without the need to restart the phone. I modified some code to prevent the Coap endpoint from setting itself as the data sink for DTLS until the session is established and the first packed buffered in DTLS. [https://source.fitbit.com/projects/XP/repos/golden-gate/pull-requests/174/overview](https://source.fitbit.com/projects/XP/repos/golden-gate/pull-requests/174/overview)

_The fix_

Pump any pending data on DTLS user side when the data sink is set and the DTLS session is established.

_QA_

*   Tested with the simulated scenario in GGHost
*   Tested the original scenario with Fitbit app by restarting the phone several times
*   Ran xp unit tests

Squashed commit of the following:

commit 31e0885346dfa4fc6bf2265885f9ae470d017687
Author: Emanuel Jarnea <ejarnea@fitbit.com>
Date:   Mon Nov 23 11:48:36 2020 +0200

    Pump any pending data on DTLS user side when the data sink is set and the a session is established.