Add platform mbedtls and link GG with it
 #### Summary of Changes
* Why do we want a platform-level MbedTLS?
    * we have a bunch of modules / components dependent on MbedTLS (for
instance, Golden Gate, Ambiq Suite and NFC). Having each on of them build
its own MbedTLS version is a waste of time, resources and the maintenance
will be tedious.
    * we need the ChaCha20 algorithm offered by MbedTLS for coredump
encryption.
    * a single library can be maintained much easier. Let's consider the
ever increasing number of devices and platforms in the future. Having the
possibility to customize each build so that it satisfies the dependencies
and features of each product / platform, might smoothen the developing
process in the future.
    * we get the latest updates (bug fixes and optimizations) by using
version v2.25.0. Currently, Golden Gate and Ambiq-Suite use v2.18, and v2.16,
respectively.
* Why don't we use the old MbedTLS (.libs/fw-v-mbedtls)? Because it's
deprecated.

* Implementation details
    * use the newer and shinier MbedTLS v2.25.0 (`.libs/fw-v-dh-mbedtls`)
    * added `CONFIG_MBEDTLS_LIB`. The flag is used to guard and compile the
library that is therefor linked with the other libraries.
    * compile all the sources in one single library (`libmbedtls.a`)
        * NOTE: The former Makefile (`.libs/fw-v-dh-mbedtls/library/Makefile`)
generates 3 separate libraries. Taking this into account, I've added
`Makefile_export` which compiles all the sources into one library.
    * feed the configuration file path to MbedTLS' Makefile so that we use
the desired configuration.

 #### Description of Testing
* Jenkins builds passed.