[iOS] IPD-182645 Identify half bond errors better
As seen in a Looker dashboard, some user encountered 200k half bond errors in a day. The errors came at a few hundred milliseconds interval, which suggests the half bond error was not identified correctly.

https://fitbit.cloud.looker.com/explore/10_advanced_communications_bq/connection_ios?qid=zV2YV95XS4UHZWJpL8kuwt&origin_space=1794&toggle=fil

Turns out, there may be a case that is not handled correctly in the app.

Squashed commit of the following:

commit 4083c2a0b6f5f496423fedfd54e480c1c3f52220
Author: Emanuel Jarnea <ejarnea@fitbit.com>
Date:   Wed Aug 4 20:48:07 2021 +0300

    Identify half bond errors better