Revert "IPD-189218 Improve link connection testability"
This reverts commit 35ed51a4e952e30710e2d1b38a68152fe0441c1b.

Why: Importing gg-xp into G3 breaks the build. The breaking happens because mbedtls has been updated only in G3 and importing gg-xp using copybara overwrites some files in G3. The solution is to revert this Pr, fix the gg-xp build for G3 in BB, import into G3 and re-add this PR.

Squashed commit of the following:

commit 4ed240b9716298bb6d84fa7e44c2d2bc74cfc4e8
Author: Emanuel Jarnea <ejarnea@fitbit.com>
Date:   Mon Feb 7 11:52:41 2022 +0200

    Revert "IPD-189218 Improve link connection testability"

    This reverts commit 35ed51a4e952e30710e2d1b38a68152fe0441c1b.