IPD-189168 Emit tls session state change event together with the tls state
In rare occasions, reading the dtls state from GGStack would result in a crash. While I was unable to reproduce the crash, it seems the issue occurs at BLE connect or disconnect and may be caused by accessing deallocated C pointers. This is likely to happen due to the delay that exists between the time when the tlsStateChange stack event is received and the time when the dtls status is read (pointers might be freed in the time between the two steps).

This solution proposes combining the two steps into a single, synchronous one. Since the dtls protocol element of the stack emits the tlsStateChange stack event, it's guaranteed the element exists when synchronously reading the dtls state from it.

Squashed commit of the following:

commit 5c62ae899f41fa99e4ccf6462eee3529886c8137
Author: Emanuel Jarnea <ejarnea@fitbit.com>
Date:   Sun Feb 13 12:29:04 2022 +0200

    Emit tls session state change event together with the tls state